mgcp: Fix compiler warning on 64bit builds
ENDPOINT_NUMBER takes the difference of two pointers. On 64bit
builds the difference is a long and the compiler then complains
about the usage of abs. We will never have thousands of endpoints
so silence the warning by casting the ENDPOINT_NUMBER to int.

mgcp_vty.c:1381:34: warning: absolute value function 'abs' given an argument of type 'long' but has parameter of
      type 'int' which may cause truncation of value [-Wabsolute-value]
                        rtp_port = rtp_calculate_port(ENDPOINT_NUMBER(endp),
                                                      ^
../../include/openbsc/mgcp_internal.h:206:31: note: expanded from macro 'ENDPOINT_NUMBER'
 #define ENDPOINT_NUMBER(endp) abs(endp - endp->tcfg->endpoints)
                              ^
mgcp_vty.c:1381:34: note: use function 'labs' instead

+++ openbsc/include/openbsc/mgcp_internal.h
@@ -203,7 +203,7 @@ struct mgcp_endpoint {
 	} osmux;
 };
 
-#define ENDPOINT_NUMBER(endp) abs(endp - endp->tcfg->endpoints)
+#define ENDPOINT_NUMBER(endp) abs((int)(endp - endp->tcfg->endpoints))
 
 struct mgcp_msg_ptr {
 	unsigned int start;
